server/requirements.txt
fastapi
uvicorn
rembg
pillow
requests
python-multipart
onnxruntime